Move AssetCompilerPass one level up
